+++ ceammc/ext/src/random/random_float.cpp
@@ -8,6 +8,8 @@ RandomFloat::RandomFloat(const PdArgs& a)
     , max_(nullptr)
     , seed_(nullptr)
 {
+    createInlet();
+    createInlet();
     createOutlet();
 
     min_ = new FloatProperty("@min", 0);
@@ -51,6 +53,24 @@ void RandomFloat::onBang()
     floatTo(0, dist(gen_.get()));
 }
 
+void RandomFloat::onInlet(size_t n, const AtomListView& lv)
+{
+    switch (n) {
+    case 1:
+
+        if (checkArgs(lv, ARG_FLOAT, ARG_FLOAT)) {
+            min_->setFloat(lv[0].asT<t_float>());
+            max_->setFloat(lv[1].asT<t_float>());
+        } else
+            min_->setList(lv);
+
+        break;
+    case 2:
+        max_->setList(lv);
+        break;
+    }
+}
+
 void setup_random_float()
 {
     ObjectFactory<RandomFloat> obj("random.float");
@@ -62,6 +82,6 @@ void setup_random_float()
     obj.setCategory("random");
     obj.setSinceVersion(0, 1);
 
-    RandomFloat::setInletsInfo(obj.classPointer(), { "bang" });
+    RandomFloat::setInletsInfo(obj.classPointer(), { "bang", "float: set min value", "float: set max value" });
     RandomFloat::setOutletsInfo(obj.classPointer(), { "float: random within \\[@min..@max) range" });
 }