+++ ceammc/ext/src/random/random_float.h
@@ -29,6 +29,7 @@ public:
     RandomFloat(const PdArgs& a);
 
     void onBang() override;
+    void onInlet(size_t n, const AtomListView& lv) override;
 };
 
 void setup_random_float();